requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-jose[cryptography]
python-multipart
pydantic-settings
pandas
numpy
xlrd==2.0.1
openpyxl
python-dateutil
pydantic
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
claude
selenium
webdriver-manager
httpx